move AnnounceAccounts out of RoutingTableView
