+++ chain/network/src/lib.rs
@@ -1,6 +1,7 @@
 pub use crate::peer_manager::peer_manager_actor::{Event, PeerManagerActor};
 
 mod accounts_data;
+mod announce_accounts;
 mod network_protocol;
 mod peer;
 mod peer_manager;
