+++ chain/network/src/routing/routing_table_view/mod.rs
@@ -1,35 +1,21 @@
 use crate::routing;
-use crate::store;
 use lru::LruCache;
-use near_primitives::network::{AnnounceAccount, PeerId};
-use near_primitives::types::AccountId;
+use near_primitives::network::PeerId;
 use parking_lot::Mutex;
-use std::collections::HashMap;
 use std::sync::Arc;
 
 #[cfg(test)]
 mod tests;
 
-const ANNOUNCE_ACCOUNT_CACHE_SIZE: usize = 10_000;
 const LAST_ROUTED_CACHE_SIZE: usize = 10_000;
 
 pub(crate) struct RoutingTableView(Mutex<Inner>);
 
 struct Inner {
-    /// Maps an account_id to a peer owning it.
-    account_peers: LruCache<AccountId, AnnounceAccount>,
-    /// Subset of account_peers, which we have broadcasted to the peers.
-    /// It is used to skip rebroadcasting the same data multiple times.
-    /// It contains less entries than account_peers in case some AnnounceAccounts
-    /// have been loaded from storage without broadcasting.
-    account_peers_broadcasted: LruCache<AccountId, AnnounceAccount>,
-
     /// For each peer, the set of neighbors which are one hop closer to `my_peer_id`.
     /// Alternatively, if we look at the set of all shortest path from `my_peer_id` to peer,
     /// this will be the set of first nodes on all such paths.
     next_hops: Arc<routing::NextHopTable>,
-    /// Access to store on disk
-    store: store::Store,
 
     /// Counter of number of calls to find_route_by_peer_id.
     find_route_calls: u64,
@@ -50,24 +36,6 @@ impl Inner {
         self.find_route_calls += 1;
         Ok(next_hop.clone())
     }
-
-    /// Get AnnounceAccount for the given AccountId.
-    fn get_announce(&mut self, account_id: &AccountId) -> Option<AnnounceAccount> {
-        if let Some(aa) = self.account_peers.get(account_id) {
-            return Some(aa.clone());
-        }
-        match self.store.get_account_announcement(&account_id) {
-            Err(e) => {
-                tracing::warn!(target: "network", "Error loading announce account from store: {:?}", e);
-                None
-            }
-            Ok(None) => None,
-            Ok(Some(a)) => {
-                self.account_peers.put(account_id.clone(), a.clone());
-                Some(a)
-            }
-        }
-    }
 }
 
 #[derive(Debug)]
@@ -77,12 +45,9 @@ pub(crate) enum FindRouteError {
 }
 
 impl RoutingTableView {
-    pub fn new(store: store::Store) -> Self {
+    pub fn new() -> Self {
         Self(Mutex::new(Inner {
-            account_peers: LruCache::new(ANNOUNCE_ACCOUNT_CACHE_SIZE),
-            account_peers_broadcasted: LruCache::new(ANNOUNCE_ACCOUNT_CACHE_SIZE),
             next_hops: Default::default(),
-            store,
             find_route_calls: 0,
             last_routed: LruCache::new(LAST_ROUTED_CACHE_SIZE),
         }))
@@ -110,69 +75,13 @@ impl RoutingTableView {
         self.0.lock().next_hops.get(peer_id).cloned()
     }
 
-    /// Find peer that owns this AccountId.
-    pub(crate) fn account_owner(&self, account_id: &AccountId) -> Option<PeerId> {
-        self.0.lock().get_announce(account_id).map(|announce_account| announce_account.peer_id)
-    }
-
-    /// Adds accounts to the routing table.
-    /// Returns the diff: new values that should be broadcasted.
-    /// Note: There is at most one peer id per account id.
-    pub(crate) fn add_accounts(&self, aas: Vec<AnnounceAccount>) -> Vec<AnnounceAccount> {
-        let mut inner = self.0.lock();
-        let mut res = vec![];
-        for aa in aas {
-            // We skip broadcasting stuff that is already broadcasted.
-            if inner.account_peers_broadcasted.get(&aa.account_id).map(|x| &x.epoch_id)
-                == Some(&aa.epoch_id)
-            {
-                continue;
-            }
-            inner.account_peers.put(aa.account_id.clone(), aa.clone());
-            inner.account_peers_broadcasted.put(aa.account_id.clone(), aa.clone());
-            // Add account to store. Best effort
-            if let Err(e) = inner.store.set_account_announcement(&aa.account_id, &aa) {
-                tracing::warn!(target: "network", "Error saving announce account to store: {:?}", e);
-            }
-            res.push(aa);
-        }
-        res
-    }
-
     pub(crate) fn info(&self) -> RoutingTableInfo {
         let inner = self.0.lock();
-        let account_peers =
-            inner.account_peers.iter().map(|(id, aa)| (id.clone(), aa.peer_id.clone())).collect();
-        RoutingTableInfo { account_peers, next_hops: inner.next_hops.clone() }
-    }
-
-    /// Public interface for `account_peers`.
-    /// Get keys currently on cache.
-    pub(crate) fn get_accounts_keys(&self) -> Vec<AccountId> {
-        self.0.lock().account_peers.iter().map(|(k, _v)| k).cloned().collect()
-    }
-
-    /// Get announce accounts on cache.
-    pub(crate) fn get_announce_accounts(&self) -> Vec<AnnounceAccount> {
-        self.0.lock().account_peers.iter().map(|(_, v)| v.clone()).collect()
-    }
-
-    /// Get AnnounceAccount for the given AccountIds, that we already broadcasted.
-    pub(crate) fn get_broadcasted_announces<'a>(
-        &'a self,
-        account_ids: impl Iterator<Item = &'a AccountId>,
-    ) -> HashMap<AccountId, AnnounceAccount> {
-        let mut inner = self.0.lock();
-        account_ids
-            .filter_map(|id| {
-                inner.account_peers_broadcasted.get(id).map(|a| (id.clone(), a.clone()))
-            })
-            .collect()
+        RoutingTableInfo { next_hops: inner.next_hops.clone() }
     }
 }
 
 #[derive(Debug)]
 pub struct RoutingTableInfo {
-    pub account_peers: HashMap<AccountId, PeerId>,
     pub next_hops: Arc<routing::NextHopTable>,
 }
